inital functionality for drag and drop
