+++ client/store/index.js
@@ -1,16 +1,16 @@
-import {createStore, combineReducers, applyMiddleware} from 'redux'
-import { createLogger } from 'redux-logger'
-import thunkMiddleware from 'redux-thunk'
-import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
-import projects from './projects'
+import { createStore, combineReducers, applyMiddleware } from 'redux';
+import { createLogger } from 'redux-logger';
+import thunkMiddleware from 'redux-thunk';
+import { composeWithDevTools } from 'redux-devtools-extension';
+import user from './user';
+import ticket from './ticket';
 
-const reducer = combineReducers({user, projects})
+const reducer = combineReducers({ user, ticket });
 const middleware = composeWithDevTools(
-  applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
-)
-const store = createStore(reducer, middleware)
+  applyMiddleware(thunkMiddleware, createLogger({ collapsed: true }))
+);
+const store = createStore(reducer, middleware);
 
-export default store
-export * from './user'
-export * from './projects'
+export default store;
+export * from './user';
+export * from './ticket';
